+++ chainerx_cc/chainerx/cuda/cuda_device/batch_norm.cc
@@ -18,6 +18,7 @@
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
 #include "chainerx/error.h"
+#include "chainerx/kernels/normalization.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/normalization.h"
